requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile setup.py
#
click==7.1.2
    # via check_dockerfile (setup.py)
dockerfile-parse==1.1.0
    # via check_dockerfile (setup.py)
pydantic==1.7.3
    # via check_dockerfile (setup.py)
pyyaml==5.3.1
    # via check_dockerfile (setup.py)
six==1.15.0
    # via dockerfile-parse
